Updated the structure
